+++ apps/web/src/components/profile/profile.css
@@ -637,3 +637,4 @@
     font-size: 0.85rem;
   }
 }
+}
